Assignment 2/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
Django==5.1.7
django-crispy-forms==2.3
django-debug-toolbar==5.1.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
idna==3.10
requests==2.32.3
setuptools==78.1.0
sqlparse==0.5.3
urllib3==2.3.0
